third_party/WebKit/LayoutTests/fast/canvas/canvas-stroke-empty-fill.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function(t) {
    var ctx = document.createElement('canvas').getContext('2d');

    ctx.fillStyle = 'rgba(0,0,0,0)';
    ctx.strokeStyle = 'green';
    ctx.lineWidth = 200;
    ctx.moveTo(0,100);
    ctx.lineTo(200,100);
    ctx.stroke();
    var imageData = ctx.getImageData(1, 0, 1, 1).data;
    assert_array_equals(imageData.slice(0, 3), [0, 128, 0]);
}, 'Verify correct behaviour for stroke() on an empty fillStyle() (alpha=0).');

</script>
